Remove NativeNavigation dependencies from CoreNavigation.
