--- MapboxCoreNavigation/SkuTokenProvider.swift
@@ -1,17 +1,0 @@
-import Foundation
-import MapboxDirections
-import MapboxNavigationNative
-
-class SkuTokenProvider: SkuTokenSource {
-    var peer: MBXPeerWrapper?
-
-    private var directionCredentials: DirectionsCredentials
-    
-    init(with credentials: DirectionsCredentials) {
-        directionCredentials = credentials
-    }
-    
-    func getToken() -> String {
-        return directionCredentials.skuToken ?? ""
-    }
-}
